requirements.txt
numpy
pandas
biopython
matplotlib
seaborn
scikit-learn
requests
pyyaml
flask
flask-cors
pytest
jupyter